Refactor view files
Reviewer: none
Issue: DEV-48
