+++ Makefile
@@ -10,7 +10,12 @@ build-wk:
 			 ./wk/command.go \
 			 ./wk/mock_command.go \
 			 ./wk/navigation.go \
-			 ./wk/views.go
+			 ./wk/views.go \
+			 ./wk/index_view.go \
+			 ./wk/summary_view.go \
+			 ./wk/assignments_view.go \
+			 ./wk/reviews_view.go \
+			 ./wk/account_view.go
 
 package-wk: build-wk
 	cd $(DIST_DIR)/wk && zip wk.zip ./wk
