+++ pkg/wanikani/assignment.go
@@ -7,6 +7,16 @@ import (
 	_ "github.com/mattn/go-sqlite3"
 )
 
+type SRSStageName string
+
+const (
+	Apprentice  SRSStageName = "apprentice"
+	Guru        SRSStageName = "guru"
+	Master      SRSStageName = "master"
+	Enlightened SRSStageName = "enlightened"
+	Burned      SRSStageName = "burned"
+)
+
 // Get assignments
 func GetAssignments(ctx context.Context, wkClient *wanikaniapi.Client) ([]*wanikaniapi.Assignment, error) {
 	isBurned := false
@@ -37,4 +47,21 @@ func GetAssignments(ctx context.Context, wkClient *wanikaniapi.Client) ([]*wanik
 	}
 
 	return assignments, nil
+}
+
+func ClassifyAssignments(assignments []*wanikaniapi.Assignment) map[SRSStageName][]*wanikaniapi.Assignment {
+	assignmentMap := make(map[SRSStageName][]*wanikaniapi.Assignment)
+	for _, assignment := range assignments {
+		switch stage := assignment.Data.SRSStage; {
+		case stage < 5:
+			assignmentMap[Apprentice] = append(assignmentMap[Apprentice], assignment)
+		case stage < 7:
+			assignmentMap[Guru] = append(assignmentMap[Guru], assignment)
+		case stage < 8:
+			assignmentMap[Master] = append(assignmentMap[Master], assignment)
+		case stage < 9:
+			assignmentMap[Enlightened] = append(assignmentMap[Enlightened], assignment)
+		}
+	}
+	return assignmentMap
 }
