+++ wk/reviews_view.go
@@ -1,0 +1,20 @@
+package main
+
+import (
+	"fmt"
+	"strings"
+)
+
+// Render reviews view
+func (m model) reviewsView() string {
+	if m.Reviews == nil {
+		return "loading..."
+	}
+	var b strings.Builder
+	b.WriteString("  Reviews:\n")
+	for i, review := range m.Reviews.Data {
+		b.WriteString(fmt.Sprintf("%d: %d\n", i, review.Data.SubjectID))
+	}
+
+	return b.String() + "\n"
+}
